requirements.txt
ttkthemes~=3.2.2
pillow~=10.4.0
pyodbc~=5.1.0
openpyxl~=3.1.5

firebase-admin~=6.5.0
protobuf~=5.28.2
fpdf~=1.7.2